Migrate regressions from 7001 to 8000
