+++ spacy/tests/pipeline/test_textcat.py
@@ -7,6 +7,7 @@ from thinc.api import Config, compounding, fix_random_seed, get_current_ops
 
 import spacy
 from spacy import util
+from spacy.cli.evaluate import print_textcats_auc_per_cat, print_prf_per_type
 from spacy.lang.en import English
 from spacy.language import Language
 from spacy.pipeline import TextCategorizer
@@ -262,6 +263,17 @@ def test_issue6908(component_name):
         config_str = config_str.replace("TRAIN_PLACEHOLDER", train_path.as_posix())
         config = util.load_config_from_str(config_str)
         init_nlp(config)
+
+
+@pytest.mark.issue(7019)
+def test_issue7019():
+    scores = {"LABEL_A": 0.39829102, "LABEL_B": 0.938298329382, "LABEL_C": None}
+    print_textcats_auc_per_cat(msg, scores)
+    scores = {
+        "LABEL_A": {"p": 0.3420302, "r": 0.3929020, "f": 0.49823928932},
+        "LABEL_B": {"p": None, "r": None, "f": None},
+    }
+    print_prf_per_type(msg, scores, name="foo", type="bar")
 
 
 @pytest.mark.skip(reason="Test is flakey when run with others")
